This development branch gets the clisp runtime and compiler working.
This development branch gets the clisp runtime and compiler working.

It implements the read function and should now generate a working
class that can read basic unsigned integers.

It is meant to get something in place that can use this compiler
branch with the runtime at:

https://github.com/jgerrish/kaitai-struct-clisp-runtime.git
or
https://gitlab.com/jgerrish/kaitai-struct-clisp-runtime.git

There is a lot of work to be done before merging, but it shows that
the generated code can use the runtime now.